% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\docType{class}
\name{xarray}
\alias{xarray}
\alias{XarrayDriver-class}
\alias{show,XarrayDriver-method}
\alias{dbUnloadDriver,XarrayDriver-method}
\alias{dbDataType,XarrayDriver-method}
\title{xarray DBI driver}
\usage{
xarray()

\S4method{show}{XarrayDriver}(object)

\S4method{dbUnloadDriver}{XarrayDriver}(drv, ...)

\S4method{dbDataType}{XarrayDriver}(dbObj, obj, ...)
}
\description{
Use with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} to open a connection to an xarray object in
a live Python session managed by reticulate.
}
\examples{
\dontrun{
con <- dbConnect(xarray(), "https://example.com/oisst.zarr", open = "open_zarr")
}
}
